Make Cookie Banner changes
Change `_cookie-banner.njk` to a macro that take parameters including
what category of cookie the banner is setting and the HTML that should
be displayed for different states of the cookie banner.

Allow `cookie-banner.mjs` to have a different cookie category than
'analytics' to be set (by using data varibles on the target module).

Move the `GDS_CONSENT_COOKIE_VERSION` to `_generic.njk` so that this
only gets set once at the top each page instead of multiple times by
different cookie banner instances.

Add new tests for multiple categories of cookie banners on the same
page. Update existing tests.