Update the guidance on when to use the checkboxes component
The guidance currently says you can use the component when you need to help users toggle a single option on or off. For these cases, it would be better to use 2 radios with yes/no or on/off because it makes users deliberately choose the option they want. Otherwise, there's the risk that users miss the single checkbox on the page and end up not selecting something they might have had selected if the page was using radio buttons.